+++ src/stores/cancelRecentPn.ts
@@ -4,4 +4,9 @@ let store: Immutable<CallStore>
 export const setCallStore = (s: Immutable<CallStore>) => {
   store = s
 }
-export const cancelRecentPn = (pnId?: string) => store.onSipUaCancel(pnId)
+export const cancelRecentPn = (n: CancelRecentPn) => store.onSipUaCancel(n)
+
+export type CancelRecentPn = {
+  pnId?: string
+  completedElseWhere?: boolean
+}